[🔥AUDIT🔥] Tokens: Export TS types correctly in package.json so consumers can use them correctly (#2564)
🖍 _This is an audit!_ 🖍

## Summary:

This change updates the `package.json` file in the `wonder-blocks-tokens`
package to ensure that TypeScript types are exported correctly. This allows
consumers of the package to use the types without any issues.

I noticed this issue when trying to use a snapshot release in Webapp. The cause
is that the `package.json`'s `field` setting was modified to support exporting
CSS files, and the change was not including the type definitions for this
package.

Issue: "none"

## Test plan:

Verify that checks pass.

In the Snapshot release, verify that the types are correctly exported and can be
used in the site properly.

Author: jandrade

Auditors: #wonder-blocks

Required Reviewers:

Approved By:

Checks: ✅ 12 checks were successful, ⏭️  2 checks have been skipped

Pull Request URL: https://github.com/Khan/wonder-blocks/pull/2564