Allow ActivityIconButton to take an arbitrary element as the icon argument (#2762)
## Summary:
I'm working on a design that adds an ActivityIconButton with Khanmigo's face as the icon in the button. This requires us to be able to pass something other than a PhosphorIcon as the `icon` prop to ActivityIconButton. This updates the prop to take an arbitrary ReactElement as well.


https://github.com/user-attachments/assets/0ed9da33-b9ea-46f3-a4ea-ef31b0e44899



## Test plan:
- Make sure tests pass
- Run storybook and check all the ActivityIconButton stories

Author: MikeKlemarewski

Reviewers: MikeKlemarewski, beaesguerra, marcysutton

Required Reviewers:

Approved By: beaesguerra

Checks: ✅ 12 checks were successful, ⏭️  2 checks have been skipped

Pull Request URL: https://github.com/Khan/wonder-blocks/pull/2762