Fix Storybook docs from rendering in rtl unexpectedly (#2787)
## Summary:

Previously, some docs pages were rendering the whole page in rtl language direction. This is because we recently updated the language direction decorator in Storybook to set the dir attribute on the document (rather than only wrapping a story), so that the `dir` could be applied to components that use portals (like the drawer component). 

To fix this, we only set the `dir` attribute on the document when the SB context `viewMode` is `story`. This makes it so it isn't applied to docs pages. 

Issue: WB-XXXX

## Test plan:
1. Confirm that the Tabs docs page (`?path=/docs/packages-tabs-tabs--docs`) renders in LTR by default. Using the Language Direction control to change to RTL should update the stories in the docs page

| Before | After | 
| --- | --- |
| <img width="1714" height="1031" alt="image" src="https://github.com/user-attachments/assets/14d32b54-e8e0-4d31-ae4d-f62ae4e70016" /> | <img width="1714" height="1028" alt="image" src="https://github.com/user-attachments/assets/1a4e6f78-6faa-4173-aef1-c3bc28abea0d" /> |

2. Confirm that the Drawer stories can still be viewed in RTL mode (`?path=/story/packages-modal-drawerlauncher-drawerlauncher--inline-start-aligned&globals=viewport:desktop;direction:rtl`)
<img width="1717" height="984" alt="image" src="https://github.com/user-attachments/assets/e2935fa9-cff1-434f-a372-b742d4819d60" />

Author: beaesguerra

Reviewers: jandrade

Required Reviewers:

Approved By: jandrade

Checks: ✅ 19 checks were successful, ⏭️  3 checks have been skipped

Pull Request URL: https://github.com/Khan/wonder-blocks/pull/2787