WB-2067: Prevent premature announcements in SingleSelect and MultiSelect (#2810)
## Summary:
To work around combobox bugs in Safari and VoiceOver, SingleSelect and MultiSelect were augmented with an Announcer integration. In practice, when a default value is set (such as "All Students" in a filter), an announcement is made on page load. The announcement is pretty random and confusing when a user isn't interacting with a dropdown. It also runs the risk of conflicting with other announcements on the page, such as client-side routing announcements or skeleton loader techniques.

In this PR, I add a flag for initial rendering that prevents the initial announcement on page load. There should still be announcements when the dropdown is opened, items filtered, and item(s) selected.

Issue: WB-2067

## Test plan:
1. Review "initial render" stories with VoiceOver turned on, listening for no announcement on page load
2. Compare to a deployed version in prod, such as the students filter on `https://classroom.khanacademy.org/teacher/class/DHJ9MS88/reports/khanmigo-activity`
3. Ensure tests pass

## Screenshots

[Before] - Storybook: "All students" filter story on WB > main branch. Announcement happens on load

<img width="563" height="366" alt="Screenshot 2025-10-01 at 4 18 01 PM" src="https://github.com/user-attachments/assets/95a74467-aad3-4fe3-9750-1bfc56b63279" />

[After] - Storybook: "All students" filter story on WB > this branch. No announcement on load.

<img width="671" height="345" alt="Screenshot 2025-10-01 at 4 19 46 PM" src="https://github.com/user-attachments/assets/ea769dba-3603-4a26-b287-a6a6e48a1699" />

[Before] - Frontend: "All students" on TX experience with announcement on load, how I originally found this issue

<img width="2079" height="1312" alt="Screenshot 2025-10-01 at 4 13 27 PM" src="https://github.com/user-attachments/assets/e1047d73-8272-4e8f-8f89-9b60637c7784" />

Note: after story for Frontend will happen when I create a snapshot integration PR from this one

Author: marcysutton

Reviewers: jandrade, marcysutton

Required Reviewers:

Approved By: jandrade

Checks: ✅ 12 checks were successful, ⏭️  2 checks have been skipped

Pull Request URL: https://github.com/Khan/wonder-blocks/pull/2810