[🔥AUDIT🔥] [hack25] Tokens: Update * as tokens imports to named imports (#2849)
🖍 _This is an audit!_ 🖍

## Summary:

Prep work to migrate the `* as tokens` imports to named imports. This is to be
able to use the upcoming codemod to migrate from `color.*` to
`semanticColor.core.*`.

Issue: "none"

## Test plan:

Verify that all the checks pass and that there are no visual regressions in
the storybook stories.

Author: jandrade

Pull Request URL: https://github.com/Khan/wonder-blocks/pull/2849