[🔥AUDIT🔥] Tooling: Update Storybook vitest config + CI workflow (#2907)
🖍 _This is an audit!_ 🖍

## Summary:

We recently moved our Storybook tests to use Vitest and included a11y tests as
well. This process is turning out to be flaky, so this PR updates the config and
the CI workflow to use their recommended setup.

https://storybook.js.org/docs/writing-tests/integrations/vitest-addon#example-configuration-files
https://storybook.js.org/docs/writing-tests/in-ci#2-add-a-new-ci-workflow

This doesn't warrant that the flakiness is fixed, so we'll need to continue to
monitor to see if it's fixed.

Here's an example of the errors we're seeing:

https://github.com/Khan/wonder-blocks/actions/runs/20183503403/job/57948995516?pr=2906#step:7:276


Issue: "none"

## Test plan:

Verify if the `Run Storybook tests` GH check passes.

Author: jandrade

Auditors: #wonder-blocks

Required Reviewers:

Approved By:

Checks: ✅ 13 checks were successful, ⏭️  3 checks have been skipped

Pull Request URL: https://github.com/Khan/wonder-blocks/pull/2907